requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.1
nltk==3.8.1
wordcloud==1.9.3
matplotlib==3.8.2
textblob==0.17.1
python-docx==1.0.1
fpdf==1.7.2
plotly==5.18.0
langdetect==1.0.9